implemented admin dictionary

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -36,8 +36,8 @@ return new class extends Migration
             $table->tinyInteger('status')->default(0)->comment('0-pending, 1-active');
             $table->integer('sequence')->default(0);
             $table->tinyInteger('public')->default(0);
-            $table->integer('readonly')->default(0);
-            $table->integer('root')->default(0);
+            $table->tinyInteger('readonly')->default(0);
+            $table->tinyInteger('root')->default(0);
             $table->tinyInteger('disabled')->default(0);
             $table->timestamps();
             $table->softDeletes();
