+++ database/migrations/0001_01_01_030000_create_admins_table.php
@@ -27,8 +27,8 @@ return new class extends Migration
             $table->string('token')->nullable();
             $table->integer('sequence')->default(0);
             $table->tinyInteger('public')->default(0);
-            $table->integer('readonly')->default(0);
-            $table->integer('root')->default(0);
+            $table->tinyInteger('readonly')->default(0);
+            $table->tinyInteger('root')->default(0);
             $table->tinyInteger('disabled')->default(0);
             $table->timestamps();
             $table->softDeletes();
