+++ database/migrations/2025_07_30_100600_create_career_skills_table.php
@@ -26,8 +26,8 @@ return new class extends Migration
             $table->string('thumbnail')->nullable();
             $table->integer('sequence')->default(0);
             $table->tinyInteger('public')->default(0);
-            $table->integer('readonly')->default(0);
-            $table->integer('root')->default(0);
+            $table->tinyInteger('readonly')->default(0);
+            $table->tinyInteger('root')->default(0);
             $table->tinyInteger('disabled')->default(0);
             $table->foreignIdFor( \App\Models\Admin::class);
             $table->timestamps();
