+++ database/migrations/2025_07_30_200200_create_career_events_table.php
@@ -22,8 +22,8 @@ return new class extends Migration
             $table->text('description');
             $table->integer('sequence')->default(0);
             $table->tinyInteger('public')->default(0);
-            $table->integer('readonly')->default(0);
-            $table->integer('root')->default(0);
+            $table->tinyInteger('readonly')->default(0);
+            $table->tinyInteger('root')->default(0);
             $table->tinyInteger('disabled')->default(0);
             $table->foreignIdFor( \App\Models\Admin::class);
             $table->timestamps();
